requirements.txt
flask
spotdl
python-dotenv
pytest
